- Move enums from Domain layer to Shared layer   - Update Budget, Transaction, RecurringTransaction and BudgetSubcategoryAllocation entities to use new enum locations   - Add AllocationType field to BudgetSubcategoryAllocation   - Update DTOs, validators and services to reflect entity changes                                                                                                                                                                                                          - Normalize email to lowercase on register and login   - Disable auto browser launch on project start/stop

+++ FinanceControl.Services/Services/AccountService.cs
@@ -1,6 +1,6 @@
 ﻿using FinanceControl.Data.Data;
 using FinanceControl.Domain.Entities;
-using FinanceControl.Domain.Enums;
+using FinanceControl.Shared.Enums;
 using FinanceControl.Domain.Interfaces.Service;
 using FinanceControl.Shared.Dtos.Others;
 using FinanceControl.Shared.Dtos.Request;
@@ -99,7 +99,7 @@ namespace FinanceControl.Services.Services
                 Id = t.Id,
                 Description = t.Description,
                 Value = t.Value,
-                Type = (int)t.Type,
+                Type = t.Type,
                 SubCategoryName = t.SubCategoryName,
                 CategoryName = t.CategoryName
             }).ToList();
